break out core spam functionality into lib
